+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-14 12:30:46
-// Commit: feat(test): integrate custom telemetry logger in test
+// Updated at: 2026-08-16 10:53:46
+// Commit: feat(test): implement shimmer loading skeleton in test
 
 class testComponentState {
-  final String commitRef = '3605';
-  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
-  final DateTime timestamp = DateTime.parse('2026-08-14T12:30:46+05:00');
+  final String commitRef = '3629';
+  final String lastMessage = 'feat(test): implement shimmer loading skeleton in test';
+  final DateTime timestamp = DateTime.parse('2026-08-16T10:53:46+05:00');
 }